Split bridge and native replay modules; unify tool completion routing.
Phase-2 maintainability follow-ups from PR #22: extract bridge server/run/snapshot modules, decompose native replay display by responsibility, and consolidate CursorSdkTurnCoordinator tool completion resolution without behavior changes.
