+++ src/cursor-native-tool-display-state.ts
@@ -1,0 +1,73 @@
+import type { CursorPiToolDisplay } from "./cursor-tool-transcript.js";
+
+export interface CursorNativeToolDisplayItem extends CursorPiToolDisplay {
+	id: string;
+	terminate?: boolean;
+}
+
+export const NATIVE_CURSOR_TOOL_DISPLAY_ENV = "PI_CURSOR_NATIVE_TOOL_DISPLAY";
+export const NATIVE_CURSOR_TOOL_REGISTRATION_ENV = "PI_CURSOR_REGISTER_NATIVE_TOOLS";
+
+export const registeredNativeToolNames = new Set<string>();
+export const nativeToolResults = new Map<string, CursorNativeToolDisplayItem>();
+
+export function readBooleanEnv(name: string): boolean | undefined {
+	const value = process.env[name]?.trim().toLowerCase();
+	if (value === "1" || value === "true" || value === "yes" || value === "on") return true;
+	if (value === "0" || value === "false" || value === "no" || value === "off") return false;
+	return undefined;
+}
+
+export function isCursorNativeToolDisplayRequested(): boolean {
+	const override = readBooleanEnv(NATIVE_CURSOR_TOOL_DISPLAY_ENV);
+	if (override !== undefined) return override;
+	return process.stdout.isTTY === true;
+}
+
+export function isCursorNativeToolRegistrationRequested(): boolean {
+	return readBooleanEnv(NATIVE_CURSOR_TOOL_REGISTRATION_ENV) !== false && isCursorNativeToolDisplayRequested();
+}
+
+export function isCursorNativeToolDisplayEnabled(): boolean {
+	return registeredNativeToolNames.size > 0;
+}
+
+export function isCursorNativeToolDisplayRuntimeEnabled(): boolean {
+	return isCursorNativeToolDisplayRequested() && registeredNativeToolNames.size > 0;
+}
+
+export function canRenderCursorToolNatively(toolName: string): boolean {
+	return registeredNativeToolNames.has(toolName);
+}
+
+export function recordCursorNativeToolDisplay(item: CursorNativeToolDisplayItem): boolean {
+	if (!canRenderCursorToolNatively(item.toolName)) return false;
+	nativeToolResults.set(item.id, item);
+	return true;
+}
+
+export function deleteCursorNativeToolDisplay(id: string): void {
+	nativeToolResults.delete(id);
+}
+
+export function consumeCursorNativeToolDisplay(id: string): CursorNativeToolDisplayItem | undefined {
+	const item = nativeToolResults.get(id);
+	if (item) nativeToolResults.delete(id);
+	return item;
+}
+
+export function isCursorReplayToolCallId(toolCallId: string): boolean {
+	return toolCallId.startsWith("cursor-replay-");
+}
+
+export function isCursorFileMutationToolName(toolName: string): toolName is "edit" | "write" {
+	return toolName === "edit" || toolName === "write";
+}
+
+export const __testUtils = {
+	nativeToolResultCount: () => nativeToolResults.size,
+	reset(): void {
+		registeredNativeToolNames.clear();
+		nativeToolResults.clear();
+	},
+};
