+++ src/cursor-pi-tool-bridge-snapshot.ts
@@ -1,0 +1,89 @@
+import type { ExtensionAPI } from "@earendil-works/pi-coding-agent";
+import { parseEnvBoolean } from "./cursor-env-boolean.js";
+import type {
+	CursorPiBridgeToolDefinition,
+	CursorPiToolBridgeSnapshot,
+	CursorPiToolBridgeSnapshotOptions,
+} from "./cursor-pi-tool-bridge-types.js";
+import { createMcpToolName, normalizeMcpInputSchema, stableNameHash } from "./cursor-pi-tool-bridge-mcp.js";
+import { isExcludedFromCursorBridgeExposure } from "./cursor-tool-names.js";
+
+export const CURSOR_PI_TOOL_BRIDGE_ENV = "PI_CURSOR_PI_TOOL_BRIDGE";
+export const CURSOR_PI_TOOL_BRIDGE_BUILTINS_ENV = "PI_CURSOR_EXPOSE_BUILTIN_TOOLS";
+
+const OVERLAPPING_CURSOR_NATIVE_PI_BUILTIN_TOOL_NAMES = new Set(["read", "bash", "write", "edit", "grep", "find", "ls"]);
+
+type CursorPiToolBridgeSnapshotApi = Pick<ExtensionAPI, "getActiveTools" | "getAllTools">;
+
+export function createEmptySnapshot(): CursorPiToolBridgeSnapshot {
+	return {
+		tools: [],
+		mcpToolNameToPiToolName: new Map(),
+		piToolNameToMcpToolName: new Map(),
+	};
+}
+
+export function resolveCursorPiToolBridgeEnabled(env: Record<string, string | undefined> = process.env): boolean {
+	return parseEnvBoolean(env[CURSOR_PI_TOOL_BRIDGE_ENV], true);
+}
+
+export function resolveCursorPiToolBridgeBuiltinsEnabled(env: Record<string, string | undefined> = process.env): boolean {
+	return parseEnvBoolean(env[CURSOR_PI_TOOL_BRIDGE_BUILTINS_ENV], false);
+}
+
+function isOverlappingCursorNativePiToolName(toolName: string): boolean {
+	return OVERLAPPING_CURSOR_NATIVE_PI_BUILTIN_TOOL_NAMES.has(toolName);
+}
+
+export function buildCursorPiToolBridgeSurfaceSignature(snapshot: CursorPiToolBridgeSnapshot): string {
+	if (snapshot.tools.length === 0) return "bridge:empty";
+	const serializedTools = snapshot.tools
+		.map((tool) =>
+			JSON.stringify({
+				piToolName: tool.piToolName,
+				mcpToolName: tool.mcpToolName,
+				description: tool.description,
+				inputSchema: tool.inputSchema,
+				source: tool.sourceInfo?.source,
+				path: tool.sourceInfo?.path,
+				scope: tool.sourceInfo?.scope,
+			}),
+		)
+		.sort()
+		.join("\0");
+	return `bridge:on:${stableNameHash(serializedTools)}`;
+}
+
+export function buildCursorPiToolBridgeSnapshot(
+	pi: CursorPiToolBridgeSnapshotApi,
+	options: CursorPiToolBridgeSnapshotOptions = {},
+): CursorPiToolBridgeSnapshot {
+	const activeToolNames = new Set(pi.getActiveTools());
+	const allTools = pi.getAllTools();
+	const usedMcpToolNames = new Set<string>();
+	const mcpToolNameToPiToolName = new Map<string, string>();
+	const piToolNameToMcpToolName = new Map<string, string>();
+	const tools: CursorPiBridgeToolDefinition[] = [];
+
+	const exposeOverlappingBuiltins = options.exposeOverlappingBuiltins === true;
+
+	for (const tool of allTools) {
+		if (!activeToolNames.has(tool.name)) continue;
+		if (isExcludedFromCursorBridgeExposure(tool.name)) continue;
+		if (!exposeOverlappingBuiltins && isOverlappingCursorNativePiToolName(tool.name)) continue;
+
+		const mcpToolName = createMcpToolName(tool.name, usedMcpToolNames);
+		const description = tool.description || `Run pi tool ${tool.name}`;
+		mcpToolNameToPiToolName.set(mcpToolName, tool.name);
+		piToolNameToMcpToolName.set(tool.name, mcpToolName);
+		tools.push({
+			piToolName: tool.name,
+			mcpToolName,
+			description,
+			inputSchema: normalizeMcpInputSchema(tool.parameters),
+			sourceInfo: tool.sourceInfo,
+		});
+	}
+
+	return { tools, mcpToolNameToPiToolName, piToolNameToMcpToolName };
+}
